requirements.txt
arrow
blinker==1.4
Bootstrap-Flask==1.0.9
certifi==2019.3.9
chardet==3.0.4
click==6.7
Faker==0.8.17
flake8==3.5.0
Flask==1.0.2
Flask-CKEditor==0.4.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itchat==1.3.10
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pynsq==0.8.3
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.7.3
python-dotenv==0.10.0
python-editor==1.0.3
PyYAML==3.13
requests==2.22.0
six==1.11.0
SQLAlchemy==1.2.10
text-unidecode==1.2
urllib3
watchdog==0.8.3
Werkzeug==0.14.1
WTForms==2.2.1
